refactor: 네이밍 변경, Overlay 구조 변경 (#167)
